App for high school
I made this app during my senior year of high school. This is an app
that is able to hold the information of all the current staff and
teachers at the school. At one time but now dismantled, it was able to
receive notifications straight from a single online server.